refactor: move DocumentEditProtocol to protocol module and inherit in CapabilityClientProtocol
