+++ src/lsp_client/protocol/__init__.py
@@ -15,7 +15,7 @@ from .capability import (
     WindowCapabilityProtocol,
     WorkspaceCapabilityProtocol,
 )
-from .client import CapabilityClientProtocol
+from .client import CapabilityClientProtocol, DocumentEditProtocol
 from .hook import (
     ServerNotificationHook,
     ServerNotificationHookExecutor,
@@ -28,6 +28,7 @@ from .hook import (
 __all__ = [
     "CapabilityClientProtocol",
     "CapabilityProtocol",
+    "DocumentEditProtocol",
     "ExperimentalCapabilityProtocol",
     "GeneralCapabilityProtocol",
     "NotebookCapabilityProtocol",
